Fix touch screen issue (touch registers as a pinch sometime)
That was happening mostly while using a big contact surface area with the screen,
for example using the thumb laying flat on the screen, or pushing hard
as it was increasing the touch value.

Might need a bit of tweaking as fingers might not be all equals ;)
It works fine with my fingers now at least.

Change-Id: I84cdb483a0dc489711038c7581737802c136179a